update imports and separate helpers and builders for views
