+++ lib/viking/view/builders/checkBoxGroupBuilder.js
@@ -1,8 +1,7 @@
-import { sanitizeToId, tagNameForModelAttribute } from '../index';
-import { checkBoxTag } from '../form_tags';
-import { label } from '../form_helpers';
+import { checkBoxTag, sanitizeToId, tagNameForModelAttribute } from '../helpers/tagHelpers';
+import { label } from '../helpers/formHelpers';
 
-export class CheckBoxGroupBuilder {
+export default class CheckBoxGroupBuilder {
 
     model;//: any
     attribute;//: string;
