+++ lib/viking/view/builders/formBuilder.js
@@ -1,6 +1,10 @@
 import Model from '../../record';
 
-import { sanitizeToId, tagNameForModelAttribute } from '../index';
+import {
+    sanitizeToId,
+    tagNameForModelAttribute
+} from '../helpers/tagHelpers';
+
 import {
     collectionSelect,
     hiddenField,
@@ -15,9 +19,9 @@ import {
     textField,
     checkBoxGroup,
     checkBox
-} from '../formHelpers';
+} from '../helpers/formHelpers';
 
-export class FormBuilder {
+export default class FormBuilder {
 
     // model: any;
     // options: any;
