+++ lib/viking/view/helpers/formHelpers.js
@@ -1,29 +1,14 @@
-import FormBuilder from 'viking/view/helpers/builders/formBuilder';
-import CheckBoxGroupBuilder from 'viking/view/helpers/builders/checkBoxGroupBuilder';
+import FormBuilder from 'viking/view/builders/formBuilder';
+import CheckBoxGroupBuilder from 'viking/view/builders/checkBoxGroupBuilder';
 
-import { humanize } from 'viking/view/support/string';
-import {
-    checkBoxTag,
-    colorFieldTag,
-    contentTag,
-    hiddenFieldTag,
-    labelTag,
-    numberFieldTag,
-    optionsForSelectTag,
-    optionsFromCollectionForSelectTag,
-    passwordFieldTag,
-    radioButtonTag,
-    selectTag,
-    textAreaTag,
-    textFieldTag
-} from 'viking/view/fromTagHelpers';
+import { humanize } from 'viking/support/string';
 
 import {
     addErrorClassToOptions,
     methodOrAttribute,
     sanitizeToId,
     tagNameForModelAttribute
-} from 'viking/view/tagHelpers';
+} from 'viking/view/helpers/tagHelpers';
 
 // checkBoxGroup(model, attribute, options = {}, content = func(f))
 // ================================================================
